Added calculation of power and trispectrum for dusty sources
